#UI- fix bugs and polish Ui

+++ admin.py
@@ -64,6 +64,12 @@ class Admin:
     def reactive_volunteer_account(self):
         print("--------------------------------------------------------------------------")
         prLightPurple("\t\t\tREACTIVATE VOLUNTEER ACCOUNT\n")
+        print(u"\U0001F531"+"Please see volunteer details below for your information: \n")  
+        conn1 = connect_db()
+        vol_df = get_volunteer_schedule_df(conn1,purpose="Status")
+        print_table(vol_df.columns,vol_df.to_numpy().tolist(),(25,25,25,25,25,25,35,35,35,35,35,35,35))
+        print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
+
         ID = Get.int(u"\U0001F539"+'Enter the volunteer ID:')
         try:
             with db.connect('emergency_system.db') as conn:
@@ -79,11 +85,6 @@ class Admin:
                 c.execute(
                     f'''UPDATE volunteer SET accountStatus = 1 WHERE volunteerID = (?)''', (ID, ))
                 conn.commit()
-                # time.sleep(1.0)
-                # vol_df = get_volunteer_schedule_df(conn,purpose="Display")
-                # sp_vol_df = vol_df.loc[vol_df["volunteerID"]==ID,"volunteerID":"accountStatus"]
-                # print("\n"+u"\U0001F538"+"Please see the volunteer status update below:\n")
-                # print_table(sp_vol_df.columns,sp_vol_df.to_numpy().tolist(),(30,30,30,30,30,30,40))
                 print("\n"+
                     u'\u2705', "Volunteer with ID {}'s account is now reactivated.".format(ID))
         except IndexError:
@@ -94,6 +95,11 @@ class Admin:
     def deactive_volunteer_account(self):
         print("--------------------------------------------------------------------------")
         prLightPurple("\t\t\tDEACTIVATE VOLUNTEER ACCOUNT\n")
+        print(u"\U0001F531"+"Please see volunteer details below for your information: \n") 
+        conn1 = connect_db()
+        vol_df = get_volunteer_schedule_df(conn1,purpose="Status")
+        print_table(vol_df.columns,vol_df.to_numpy().tolist(),(25,25,25,25,25,25,35,35,35,35,35,35,35))
+        print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
         ID = Get.int(u"\U0001F539"+'Enter the volunteer ID:')
         try:
             with db.connect('emergency_system.db') as conn:
@@ -106,17 +112,26 @@ class Admin:
                 warn("This account is currently inactive.")
                 print("-------------------------------------")
             else:
-                if confirm(f"Do you want to deactivate this account ID {ID}?"):
-                    c.execute(
+                if confirm(u"\U0001F539"+f"Do you want to deactivate this account ID {ID}?"):
+                    day_col = list(vol_df.columns[6:])
+                    vol_df_byID = vol_df.loc[vol_df["volunteerID"]==ID,:]
+                    has_req = False
+                    for d in day_col:
+                        for ind in vol_df_byID.index:
+                            if vol_df_byID[d][ind] == u"\U0001F4D1":
+                                has_req = True
+                    if has_req == False:
+                        c.execute(
                         f'''UPDATE volunteer SET accountStatus = 0 WHERE volunteerID = (?)''', (ID,))
-                    conn.commit()
-                    # time.sleep(2.0)
-                    # vol_df = get_volunteer_schedule_df(conn,purpose="Display")
-                    # sp_vol_df = vol_df.loc[vol_df["volunteerID"]==ID,"volunteerID":"accountStatus"]
-                    # print("\n"+u"\U0001F538"+"Please see the volunteer status update below:\n")
-                    # print_table(sp_vol_df.columns,sp_vol_df.to_numpy().tolist(),(30,30,30,30,30,30,40))
-                    print("\n"+
+                        conn.commit()
+                        print("\n"+
                         u'\u2705', "Volunteer with ID {}'s account is now deactive".format(ID))
+                    else:
+                        print("\n")
+                        print_table(vol_df_byID.columns,vol_df_byID.to_numpy().tolist(),(25,25,25,25,25,25,35,35,35,35,35,35,35))
+                        print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
+                        warn("You cannot deactivate this volunteer account because there is still a task request from the refugee.")
+                        return
                 else:
                     return
         except IndexError:
@@ -277,6 +292,7 @@ class Admin:
                 print_table(fd.columns,fd.to_numpy().tolist(),(12,20,20,18,12,15))
                 prCyan("\n"+u"\U0001F538"+"---VOLUNTEER AVAILABILITY SCHEDULE---\n")
                 print_table(sch.columns,sch.to_numpy().tolist(),(12,15,15,15,15,18,18,18,18,18,18,18))
+                print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
         except IndexError:
             print_log("{} is an invalid ID".format(ID))
         except:
@@ -302,6 +318,7 @@ class Admin:
                 print_table(fd.columns,fd.to_numpy().tolist(),(12,20,20,18,12,15))
                 prCyan("\n"+u"\U0001F538"+"---VOLUNTEER AVAILABILITY SCHEDULE---\n")
                 print_table(sch.columns,sch.to_numpy().tolist(),(12,15,15,15,15,18,18,18,18,18,18,18))
+                print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
         except:
             print_log("Wrong connection to the database.")
 
@@ -324,7 +341,7 @@ class Admin:
                 print_table(fd.columns,fd.to_numpy().tolist(),(12,20,20,18,12,15))
                 prCyan("\n"+u"\U0001F538"+"---VOLUNTEER AVAILABILITY SCHEDULE---\n")
                 print_table(sch.columns,sch.to_numpy().tolist(),(12,15,15,15,15,18,18,18,18,18,18,18))
-                
+                print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")       
         except:
             print_log("Wrong connection to the database")
 
@@ -332,7 +349,12 @@ class Admin:
     def delete_account(self):
         print("--------------------------------------------------------------------------")
         prLightPurple("\t\t\tDELETE VOLUNTEER ACCOUNT\n")
-        ID = input(u"\U0001F539"+'Enter the volunteer ID you would like to delete:')
+        print(u"\U0001F531"+"Please see volunteer details below for your information: \n") 
+        conn1 = connect_db()
+        vol_df = get_volunteer_schedule_df(conn1,purpose="Status")
+        print_table(vol_df.columns,vol_df.to_numpy().tolist(),(25,25,25,25,25,25,35,35,35,35,35,35,35))
+        print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
+        ID = Get.int(u"\U0001F539"+'Enter the volunteer ID you would like to delete:')
         print("\n")
         try:
             with db.connect('emergency_system.db') as conn:
@@ -347,13 +369,28 @@ class Admin:
                     print_table(fd.columns,fd.to_numpy().tolist(),(12,20,20,20,15,18))
                     confirm = AccountCreation.confirm_deletion()
                     if confirm == 1:
-                        c = conn.cursor()
-                        sql = '''DELETE FROM volunteer WHERE volunteerID = (?)'''
-                        c.execute(sql, (ID,))
-                        conn.commit()
-                        c.execute(f"insert into deleted_vol_account (volunteerID, username, password) "
+                        day_col = list(vol_df.columns[6:])
+                        vol_df_byID = vol_df.loc[vol_df["volunteerID"]==ID,:]
+                        has_req = False
+                        for d in day_col:
+                            for ind in vol_df_byID.index:
+                                if vol_df_byID[d][ind] == u"\U0001F4D1":
+                                    has_req = True
+                    
+                        if has_req == False:
+                            c = conn.cursor()
+                            sql = '''DELETE FROM volunteer WHERE volunteerID = (?)'''
+                            c.execute(sql, (ID,))
+                            conn.commit()
+                            c.execute(f"insert into deleted_vol_account (volunteerID, username, password)"
                                   f"values({ID}, '{a[0][3]}', '{a[0][-1]}') ")
-                        print(u'\u2705'+'The account is successfully deleted.')
+                            print(u'\u2705'+'The account is successfully deleted.')
+                        else:
+                            print("\n")
+                            print_table(vol_df_byID.columns,vol_df_byID.to_numpy().tolist(),(25,25,25,25,25,25,35,35,35,35,35,35,35))
+                            print("\nNote:"+u"\U00002705"+" = Free, "+u"\U0000274C"+" = Unavailable,"+u"\U0001F4D1"+" = Booked \n")
+                            warn("You cannot delete this volunteer account because there is still a task request from the refugee.")
+                            return
                 else:
                     raise IndexError
